+++ app/models/response.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 
 from datetime import datetime
 
-from sqlalchemy import DateTime, Integer, ForeignKey
+from sqlalchemy import BigInteger, DateTime, Integer, ForeignKey
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.models.base import Base
@@ -19,5 +19,5 @@ class Response(Base):
     id: Mapped[int] = mapped_column(Integer, primary_key=True)
     order_id: Mapped[int] = mapped_column(Integer, ForeignKey("orders.id"))
     # Stores master Telegram ID (not users.id).
-    master_id: Mapped[int] = mapped_column(Integer, ForeignKey("users.telegram_id"))
+    master_id: Mapped[int] = mapped_column(BigInteger, ForeignKey("users.telegram_id"))
     response_time: Mapped[datetime] = mapped_column(DateTime, default=datetime.utcnow)
